changed casting of input from int to float

+++ temperatureConversion.py
@@ -1,2 +1,2 @@
-celsius = int(input("Please enter the temperature in celsius: "))
+celsius = float(input("Please enter the temperature in celsius: "))
 print("temperature in fahrenheit: " + str((celsius *9/5) + 32) + "F")
